#228 Refactored plugins of InventoryCatalog module
