+++ app/code/Magento/InventoryCatalog/Plugins/Inventory/Stock/PreventDeletingStockWithSaleChannelAssignedPlugin.php
@@ -5,7 +5,7 @@
  */
 declare(strict_types=1);
 
-namespace Magento\InventoryCatalog\Plugin\InventoryApi\StockRepository\PreventDeleting;
+namespace Magento\InventoryCatalog\Plugins\Inventory\Stock;
 
 use Magento\Framework\Exception\CouldNotDeleteException;
 use Magento\InventoryApi\Api\StockRepositoryInterface;
@@ -14,7 +14,7 @@ use Magento\InventorySales\Model\GetAssignedSalesChannelsForStockInterface;
 /**
  * Prevent deleting of Stock which assigned at least one Sales Channel
  */
-class AssignedToSalesChannelsStockPlugin
+class PreventDeletingStockWithSaleChannelAssignedPlugin
 {
     /**
      * @var GetAssignedSalesChannelsForStockInterface
@@ -35,13 +35,17 @@ class AssignedToSalesChannelsStockPlugin
      *
      * @param StockRepositoryInterface $subject
      * @param int $stockId
+     *
      * @return void
+     *
      * @throws CouldNotDeleteException
+     *
      * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
     public function beforeDeleteById(StockRepositoryInterface $subject, int $stockId)
     {
         $assignSalesChannels = $this->assignedSalesChannelsForStock->execute($stockId);
+
         if (count($assignSalesChannels)) {
             throw new CouldNotDeleteException(__('Stock has at least one sale channel and could not be deleted.'));
         }
