+++ app/code/Magento/InventoryCatalog/etc/di.xml
@@ -9,12 +9,12 @@
     <preference for="Magento\InventoryCatalog\Api\DefaultStockProviderInterface" type="Magento\InventoryCatalog\Model\DefaultStockProvider"/>
     <preference for="Magento\InventoryCatalog\Api\DefaultSourceProviderInterface" type="Magento\InventoryCatalog\Model\DefaultSourceProvider"/>
     <type name="Magento\InventoryApi\Api\StockRepositoryInterface">
-        <plugin name="prevent_default_stock_deleting"
-                type="Magento\InventoryCatalog\Plugin\InventoryApi\StockRepository\PreventDeleting\DefaultStockPlugin"/>
-        <plugin name="prevent_assigned_to_sales_channels_stock_deleting"
-                type="Magento\InventoryCatalog\Plugin\InventoryApi\StockRepository\PreventDeleting\AssignedToSalesChannelsStockPlugin"/>
+        <plugin name="prevent_deleting_default_stock"
+                type="Magento\InventoryCatalog\Plugin\Inventory\Stock\PreventDeletingDefaultStockPlugin"/>
+        <plugin name="prevent_deleting_stock_assigned_to_at_least_one_sales_channel"
+                type="Magento\InventoryCatalog\Plugin\Inventory\Stock\PreventDeletingStockWithSaleChannelAssignedPlugin"/>
     </type>
     <type name="Magento\InventoryApi\Api\ReservationsAppendInterface">
-        <plugin name="update_legacy_catalog_inventory_plugin" type="Magento\InventoryCatalog\Plugin\Model\UpdateLegacyCatalogInventoryPlugin"/>
+        <plugin name="update_legacy_catalog_inventory_plugin" type="Magento\InventoryCatalog\Plugin\Inventory\Reservation\UpdateLegacyCatalogInventoryPlugin"/>
     </type>
 </config>
